♻️ NWVCAG code: split workflow authority boundaries
